deployment/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
sqlite3
requests==2.31.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
textblob==0.17.1
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2